Refactor tests to have one test suite per endpoint instead of individual tests per verb
